fix(tcp): 时钟锁存触发+定时重标+僵尸槽清理 — P2 三真bug
P2 取证(板测+pcap+JTAG 三重证据)定位的三个真实固件缺陷:

1. 固件时钟恒0: local_time_l(0x07) 是写0x06拍照式锁存器, 全固件无人触发
   → ISN 恒0(pcap: SYN-ACK seq 0), TCP 全部定时器死转, TIME_WAIT/idle
   永不过期, 16 连接槽被僵尸占满后新 SYN 被静默丢弃(历史'空响应/板子越用越死')
   修复: LCPU_LOCAL_TIME64() 先写0x06触发锁存再读0x07/0x08拼64位

2. 定时常数 20.6x 错标: 常数按 50MHz 标定, 板测实测自由计数器 1.035GHz
   (jpoll 采样, 32位 4.15s 回卷, latch_delta 负值抓获)
   修复: 常数重标 + 计时数组 uint64_t

3. LAST_ACK 僵尸槽: 板卡 FIN|ACK 从未上线且 final-ACK 精确匹配过苛,
   PC 端口复用后新 SYN 被僵尸吞掉(pcap 全程抓获 0.1s档#11)
   修复: 收 FIN 后 RST|ACK 立即关闭

附带: P0 取证计数器(debug_rw_0/1/3: 弹包数/解析数/push数/包首字),
本次板卡清白判定的直接功臣, 保留为长期诊断钩子。

板测: 密度矩阵 2s/0.5s/0.1s 全过, 25连接@0.2s 压力 25/25(旧固件必死)。

+++ c/inc/tcp.h
@@ -88,9 +88,9 @@ extern uint32_t connection_dst_ips[MAX_CONNECTIONS];
 extern uint32_t available_connections;
 
 // Timer / housekeeping fields
-extern uint32_t connection_time_wait_start[MAX_CONNECTIONS];
-extern uint32_t connection_last_activity[MAX_CONNECTIONS];
-extern uint32_t connection_last_tx_time[MAX_CONNECTIONS];
+extern uint64_t connection_time_wait_start[MAX_CONNECTIONS];
+extern uint64_t connection_last_activity[MAX_CONNECTIONS];
+extern uint64_t connection_last_tx_time[MAX_CONNECTIONS];
 extern uint8_t  connection_syn_retries[MAX_CONNECTIONS];
 
 void tcp_connection_init();
